Improve text rendering capability of wasm and simplify code
