+++ crates/monocurl/Cargo.toml
@@ -28,7 +28,7 @@ executor.workspace = true
 exporter.workspace = true
 geo.workspace = true
 latex.workspace = true
-stdlib = { workspace = true, features = ["native-latex", "native-tessellation"] }
+stdlib.workspace = true
 structs.workspace = true
 bytecode.workspace = true
