+++ crates/runtime/src/lib.rs
@@ -106,6 +106,7 @@ pub struct ExecutionSnapshot {
     pub camera: Option<CameraSnapshot>,
     pub camera_version: Option<u64>,
     pub meshes: Option<Vec<Arc<Mesh>>>,
+    pub errors: Vec<executor::error::RuntimeError>,
     pub current_timestamp: Timestamp,
     pub target_timestamp: Timestamp,
     pub status: ExecutionStatus,
@@ -659,6 +660,7 @@ fn runtime_snapshot(
         camera,
         camera_version,
         meshes,
+        errors: executor.state.errors.clone(),
         current_timestamp,
         target_timestamp,
         status,
